pyproject.toml
[project]
name = "paperclips"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-limiter>=3.10.1",
    "openai>=1.61.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.9.4",
]
